refactor: move getFormattedNumberString to utils
